+++ test/fuzz/SellOfferClamp.t.sol
@@ -180,7 +180,7 @@ contract SellOfferClampFuzzTest is ClampFuzzFixtures {
         bytes32 group = keccak256(abi.encodePacked("shares", offerCapacity, tick));
 
         // SELL offer: maker=borrower (seller), taker=lender (buyer)
-        (uint256 mu, uint256 ma) = _denomFields(offerCapacity, denom);
+        (uint128 mu, uint128 ma) = _denomFields(offerCapacity, denom);
         Offer memory offer = Offer({
             market: market,
             buy: false,
@@ -254,7 +254,7 @@ contract SellOfferClampFuzzTest is ClampFuzzFixtures {
 
         bytes32 group = keccak256(abi.encodePacked("units", offerCapacity, tick));
 
-        (uint256 mu, uint256 ma) = _denomFields(offerCapacity, denom);
+        (uint128 mu, uint128 ma) = _denomFields(offerCapacity, denom);
         Offer memory offer = Offer({
             market: market,
             buy: false,
@@ -342,7 +342,7 @@ contract SellOfferClampFuzzTest is ClampFuzzFixtures {
         midnight.setIsAuthorized(address(ecrecoverRatifier), true, freshBorrower);
         vm.stopPrank();
 
-        (uint256 mu, uint256 ma) = _denomFields(offerCapacity, denom);
+        (uint128 mu, uint128 ma) = _denomFields(offerCapacity, denom);
         Offer memory offer = Offer({
             market: market,
             buy: false,
@@ -434,7 +434,7 @@ contract SellOfferClampFuzzTest is ClampFuzzFixtures {
         vm.prank(buyer2);
         loanToken.approve(address(midnight), type(uint256).max);
 
-        (uint256 mu, uint256 ma) = _denomFields(offerCapacity, denom);
+        (uint128 mu, uint128 ma) = _denomFields(offerCapacity, denom);
         Offer memory offer = Offer({
             market: market,
             buy: false,
